+++ src/ntfs2btrfs.h
@@ -251,4 +251,4 @@ void read_nonresident_mappings(const ATTRIBUTE_RECORD_HEADER* att, std::list<map
                                uint32_t cluster_size);
 std::string_view find_sd(uint32_t id, ntfs_file& secure, ntfs& dev);
 void populate_skip_list(ntfs& dev, uint64_t inode, std::list<uint64_t>& skiplist);
-void process_fixups(MULTI_SECTOR_HEADER* header, unsigned int length, unsigned int sector_size);
+void process_fixups(MULTI_SECTOR_HEADER* header, uint64_t length, unsigned int sector_size);
